Assignment 4.1 NPTEL Array Sum Given an array A of N positive integers, you are provided with two functions sumprefix(i) and sumsuffix(i). The sumprefix(i) function computes the sum of first i numbers of the array. The sumsuffix(i) function computes the sum of last N - i + 1 numbers of the array. Your task is to find the minimum index i for which the value sumprefix(i) + sumsuffix(i) is the minimum. Which means you need to minimize the value of sumprefix(i) + sumsuffix(i) and find the least index i for which this value is attained.
Input:
The first line of input contains the integer N denoting the number of elements in array A.
The second line contains the N separated elements of array A.

Output:
Output the single line containing the answer.

Constraints:
1 <= N, A[i] <= 105

Example:
Input:
3
1 2 3

Output:
1

Explanation:
Let's calculate the sumprefix(i) + sumsuffix(i) for all i.

sumprefix(1) + sumsuffix(1) = 1 + 6 = 7
sumprefix(2) + sumsuffix(2) = 3 + 5 = 8
sumprefix(3) + sumsuffix(3) = 6 + 3 = 9

The minimum value of the function is 7 which is attained at index 1, so the answer is 1.